requirements.txt
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

anthropic==0.34.1

python-dotenv==1.0.1

httplib2==0.22.0
requests==2.31.0

tqdm==4.66.4